requirements.txt
# Core dependencies
playwright>=1.49.0
pydantic>=2.9.0
pyyaml==6.0.1

# Scheduling (optional)
schedule==1.2.0
croniter==2.0.1

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development tools (optional)
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Documentation (optional)
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
